Abolish rotate
